backend/python_services/summarizer/requirements.txt
firebase-admin>=6.5.0
langchain>=0.2.0
langgraph>=0.0.48
langchain-google-genai>=1.0.3
python-dotenv>=1.0.1